+++ app/renderer/pages/wallet/account/AccountDetails.js
@@ -3,18 +3,25 @@ import R from 'ramda';
 import styled from 'styled-components';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { MainContent, MainLayout } from 'components/layout';
-import { PageHeading, PageFooter, Tabs, TabPane } from 'components';
+import { Clipboard, PageHeading, PageFooter, Tabs, TabPane } from 'components';
 import PortfolioSection from './PortfolioSection';
 import ReceiveSection from './ReceiveSection';
 
-const AccountDetails = ({ account }) => {
+const Subheading = ({ account }) => {
   return (
+    <div style={{ display: 'flex' }}>
+      <span>Public Address: </span>
+      <Clipboard>{account.address}</Clipboard>
+    </div>
+  );
+};
+
+const AccountDetails = ({ account }) => {
+  return account ? (
     <React.Fragment>
-      <PageHeading subHeading={`Public Address: ${account.address}`}>
-        Account name here...
-      </PageHeading>
+      <PageHeading subHeading={<Subheading {...{ account }} />}>Account name here...</PageHeading>
       <div className="content">
-        <Tabs defaultActiveKey="1" styles={{ maxHeight: '730px' }}>
+        <Tabs defaultActiveKey="2" styles={{ maxHeight: '730px' }}>
           <TabPane tab="Portfolio" key="1">
             <PortfolioSection {...{ account }} />
           </TabPane>
@@ -25,6 +32,8 @@ const AccountDetails = ({ account }) => {
       </div>
       <PageFooter blank />
     </React.Fragment>
+  ) : (
+    <PageHeading subHeading="There is no accessible account"> Wallet account </PageHeading>
   );
 };