+++ app/renderer/pages/wallet/rootPage/container.js
@@ -16,9 +16,10 @@ const enhance = compose(
     componentDidMount() {
       const storedWallets = this.props.localStorage[storageKeys.WALLETS];
       const hasWallet = storedWallets && storedWallets.length > 0;
+      const firstWallet = hasWallet && storedWallets[0];
 
       const nextRoute = hasWallet
-        ? `${ROUTES.WALLET.ROOT}/${storedWallets[0].id}/accounts/${Object.keys(storedWallets[0].accounts)[0]}`
+        ? `${ROUTES.WALLET.ROOT}/${firstWallet.id}/accounts/${Object.keys(firstWallet.accounts)[0]}`
         : ROUTES.WALLET.LANDING;
 
       history.push(nextRoute);